% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elConnect.R
\name{elConnect}
\alias{elConnect}
\title{A helper function to connect to the database using .dbconf}
\usage{
elConnect(config = "client")
}
\arguments{
\item{config}{Will use this config in the .dbconf file. Defaual 'client'}
}
\value{
A connection object as returned from RMySQL
}
\description{
A helper function to connect to the database using .dbconf
}
\examples{
con = dbConnect()

}
\seealso{
\code{link[RMySQL]}
}
